# docs/cs/openapi.yaml
openapi: 3.1.0
info:
  title: 2Africa CS-Core API
  version: 1.0.0-rc.1
  summary: Horizontal customer-service module shared across the 2Africa product family.
  description: |
    The CS-Core API is the read+write surface for the cross-product
    Customer Service module embedded in AgriOS / RetailOS / FactoryOS /
    TravelOS / AgriCloud / MarketOS. It is the same wire format whether
    you call it via `/v1/cs/*` (canonical) or `/v1/service/*` (legacy
    alias, kept for backward compatibility through v4.x).

    See `docs/cs/README.md` for the architectural overview and
    `docs/cs/INTEGRATION.md` for how consuming products plug in their
    own `BusinessContextProvider`.

  contact:
    name: 2Africa.AI
    url: https://github.com/2AfricaAI/2Africa-AgriOS

  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080/api
    description: Local dev (embedded in AgriOS)
  - url: http://localhost:4410/api
    description: Future standalone ServiceOS (see docs/cs/PORTS.md)
security:
  - bearerAuth: []
...
paths:
...
  /v1/cs/inboxes/setup-email:
    post:
      tags: [Channels]
      summary: Provision an Email inbox (IMAP + SMTP)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailInboxSetup'
      responses:
        '200':
          description: Created inbox.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeInbox'
  /v1/cs/inboxes/setup-whatsapp:
    post:
      tags: [Channels]
      summary: Provision a WhatsApp Cloud inbox (Meta Business API)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WhatsAppInboxSetup'
      responses:
        '200':
          description: Created inbox.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeInbox'
  /v1/cs/inboxes/setup-web-widget:
    post:
      tags: [Channels]
      summary: Provision a Web Widget inbox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebWidgetInboxSetup'
      responses:
        '200':
          description: Created inbox.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeInbox'
  /v1/cs/inboxes/setup-sms:
    post:
      tags: [Channels]
      summary: Provision an SMS inbox (Africa's Talking / Twilio-compatible API channel)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmsInboxSetup'
      responses:
        '200':
          description: Created inbox + inbound webhook URL the operator must register with the SMS provider.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeSmsInboxSetupResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    EnvelopeMeta:
      type: object
      properties:
        code: { type: integer, example: 200 }
        msg: { type: string, example: OK }
        traceId: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
      required: [code, msg]
...
    EnvelopeInbox:
      allOf:
        - $ref: '#/components/schemas/EnvelopeMeta'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Inbox' }
...
    EnvelopeSmsInboxSetupResult:
      allOf:
        - $ref: '#/components/schemas/EnvelopeMeta'
        - type: object
          properties:
            data:
              type: object
              properties:
                inbox: { $ref: '#/components/schemas/Inbox' }
                inboundWebhookUrl: { type: string, description: 'Register this URL with the SMS provider so they POST inbound SMS into CS-Core.' }

    # --- Domain resources ---
...
    Inbox:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        channelType: { type: string, description: 'Channel::WebWidget / Channel::Email / Channel::Whatsapp / Channel::Api etc.' }
        email: { type: string, nullable: true, format: email }
        phoneNumber: { type: string, nullable: true }
...
    EmailInboxSetup:
      type: object
      required: [name, email]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        imapAddress: { type: string, nullable: true }
        imapPort: { type: integer, nullable: true }
        smtpAddress: { type: string, nullable: true }
        smtpPort: { type: integer, nullable: true }
        appPassword: { type: string, description: 'Gmail app password preferred over OAuth for the open-box flow.' }
    WhatsAppInboxSetup:
      type: object
      required: [name, phoneNumberId, accessToken, businessAccountId]
      properties:
        name: { type: string }
        phoneNumberId: { type: string }
        accessToken: { type: string }
        businessAccountId: { type: string }
        webhookVerifyToken: { type: string, nullable: true }
    WebWidgetInboxSetup:
      type: object
      required: [name, websiteUrl]
      properties:
        name: { type: string }
        websiteUrl: { type: string, format: uri }
        welcomeTitle: { type: string, nullable: true }
        welcomeTagline: { type: string, nullable: true }
        widgetColor: { type: string, nullable: true, description: 'Hex like #2BA84A' }
    SmsInboxSetup:
      type: object
      required: [name, provider, phoneNumber]
      properties:
        name: { type: string }
        provider: { type: string, enum: [africastalking, twilio] }
        phoneNumber: { type: string }
        senderId: { type: string, nullable: true }

    # --- RFC 7807 problem+json ---
...